ER Historial Familiar CRUD
Añadidas validaciones por clinica de historial familiar

+++ src/Controller/HistorialFamiliarController.php
@@ -40,28 +40,12 @@ class HistorialFamiliarController extends AbstractController
 
         if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
             if($AuthUser->getUser()->getClinica()->getId() == $familiar->getFamiliaresExpedientes()[0]->getExpediente()->getUsuario()->getClinica()->getId()){
-                if($familiar->getFamiliaresExpedientes()[0]->getExpediente()->getHabilitado()){
-                    $em = $this->getDoctrine()->getManager();
-                    $RAW_QUERY = "SELECT historial.* FROM historial_familiar as historial, familiar as fam WHERE fam.id=historial.familiar_id AND familiar_id=".$familiar->getId().";";
-                    $statement = $em->getConnection()->prepare($RAW_QUERY);
-                    $statement->execute();
-                    $result = $statement->fetchAll();
-                    return $this->render('historial_familiar/index.html.twig', [
-                        'historial_familiares' => $result,
-                        'user'              => $AuthUser,
-                        'familiar'          => $familiar,
-                        'expediente'        => $expediente
-                    ]);
-                }else{
-                    $this->addFlash('fail','Este paciente no está habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
-                    return $this->redirectToRoute('expediente_index');
-                }
             }else{
                 $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
                 return $this->redirectToRoute('expediente_index');
             }  
         } 
-        
+
         if($familiar->getFamiliaresExpedientes()[0]->getExpediente()->getHabilitado()){
             $em = $this->getDoctrine()->getManager();
             $RAW_QUERY = "SELECT historial.* FROM historial_familiar as historial, familiar as fam WHERE fam.id=historial.familiar_id AND familiar_id=".$familiar->getId().";";
@@ -90,42 +74,6 @@ class HistorialFamiliarController extends AbstractController
     {
         if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
             if($AuthUser->getUser()->getClinica()->getId() == $familiar->getFamiliaresExpedientes()[0]->getExpediente()->getUsuario()->getClinica()->getId()){
-                if($familiar->getFamiliaresExpedientes()[0]->getExpediente()->getHabilitado()){
-                    $editar = false;
-                    $historialFamiliar = new HistorialFamiliar();
-                    $form = $this->createForm(HistorialFamiliarType::class, $historialFamiliar);
-                    $form->handleRequest($request);
-
-                    if ($form->isSubmitted() && $form->isValid()) {
-                        $entityManager = $this->getDoctrine()->getManager();
-                        if($form["descripcion"]->getData() != ""){
-                            $historialFamiliar->setDescripcion($form["descripcion"]->getData());
-                            $historialFamiliar->setFamiliar($familiar);
-                            $entityManager->persist($historialFamiliar);
-                            $entityManager->flush();
-                        }else{
-                            $this->addFlash('fail', 'Error, el campo de la descripción no puede ir vacía');
-                            return $this->render('historial_familiar/new.html.twig', [
-                                'historial_familiar'  => $historialFamiliar,
-                                'familiar'          => $familiar,
-                                'editar'            => $editar,
-                                'form'              => $form->createView(),
-                            ]);
-                        }
-                        $this->addFlash('success', 'Historial añadido con éxito');
-                        return $this->redirectToRoute('historial_familiar_index',['familiar' => $familiar->getId()]);
-                    }
-
-                    return $this->render('historial_familiar/new.html.twig', [
-                        'historial_familiar' => $historialFamiliar,
-                        'familiar' => $familiar,
-                        'editar'     => $editar,
-                        'form' => $form->createView(),
-                    ]);
-                }else{
-                    $this->addFlash('fail','Este paciente no está habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
-                    return $this->redirectToRoute('expediente_index');
-                }
             }else{
                 $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
                 return $this->redirectToRoute('expediente_index');
@@ -177,12 +125,26 @@ class HistorialFamiliarController extends AbstractController
      * @Security2("user.getIsActive()", statusCode=412, message="Su cuenta está inactiva")
      * @Security2("is_granted('ROLE_PERMISSION_SHOW_HISTORIAL_FAMILIAR')")
      */
-    public function show(HistorialFamiliar $historialFamiliar, Familiar $familiar): Response
+    public function show(HistorialFamiliar $historialFamiliar, Familiar $familiar, Security $AuthUser): Response
     {
-        return $this->render('historial_familiar/show.html.twig', [
-            'historial_familiar' => $historialFamiliar,
-            'familiar'       => $familiar,
-        ]);
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $familiar->getFamiliaresExpedientes()[0]->getExpediente()->getUsuario()->getClinica()->getId() && $familiar->getFamiliaresExpedientes()[0]->getFamiliar()->getId() == $historialFamiliar->getFamiliar()->getId() ){
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('expediente_index');
+            }  
+        }
+
+        if($familiar->getFamiliaresExpedientes()[0]->getExpediente()->getHabilitado()){
+            return $this->render('historial_familiar/show.html.twig', [
+                'historial_familiar' => $historialFamiliar,
+                'familiar'       => $familiar,
+            ]);
+        }else{
+            $this->addFlash('fail','Este paciente no está habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+            return $this->redirectToRoute('expediente_index');
+        }
+
     }
 
     /**
@@ -193,22 +155,35 @@ class HistorialFamiliarController extends AbstractController
      */
     public function edit(Request $request, HistorialFamiliar $historialFamiliar, Familiar $familiar): Response
     {
-        $editar = true;
-        $form = $this->createForm(HistorialFamiliarType::class, $historialFamiliar);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-            $this->addFlash('success', 'Historial modificado con éxito');
-            return $this->redirectToRoute('historial_familiar_index',['familiar' => $familiar->getId()]);
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $familiar->getFamiliaresExpedientes()[0]->getExpediente()->getUsuario()->getClinica()->getId() && $familiar->getFamiliaresExpedientes()[0]->getFamiliar()->getId() == $historialFamiliar->getFamiliar()->getId() ){
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('expediente_index');
+            }  
         }
 
-        return $this->render('historial_familiar/edit.html.twig', [
-            'historial_familiar' => $historialFamiliar,
-            'familiar'       => $familiar,
-            'editar'     => $editar,
-            'form' => $form->createView(),
-        ]);
+        if($familiar->getFamiliaresExpedientes()[0]->getExpediente()->getHabilitado()){
+            $editar = true;
+            $form = $this->createForm(HistorialFamiliarType::class, $historialFamiliar);
+            $form->handleRequest($request);
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+                $this->addFlash('success', 'Historial modificado con éxito');
+                return $this->redirectToRoute('historial_familiar_index',['familiar' => $familiar->getId()]);
+            }
+
+            return $this->render('historial_familiar/edit.html.twig', [
+                'historial_familiar' => $historialFamiliar,
+                'familiar'       => $familiar,
+                'editar'     => $editar,
+                'form' => $form->createView(),
+            ]);
+        }else{
+            $this->addFlash('fail','Este paciente no está habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+            return $this->redirectToRoute('expediente_index');
+        }
     }
 
     /**
@@ -219,13 +194,26 @@ class HistorialFamiliarController extends AbstractController
      */
     public function delete(Request $request, HistorialFamiliar $historialFamiliar, Familiar $familiar): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$historialFamiliar->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($historialFamiliar);
-            $entityManager->flush();
+        if($AuthUser->getUser()->getRol()->getNombreRol() != 'ROLE_SA'){
+            if($AuthUser->getUser()->getClinica()->getId() == $familiar->getFamiliaresExpedientes()[0]->getExpediente()->getUsuario()->getClinica()->getId() && $familiar->getFamiliaresExpedientes()[0]->getFamiliar()->getId() == $historialFamiliar->getFamiliar()->getId() ){
+            }else{
+                $this->addFlash('fail','Error, este registro puede que no exista o no le pertenece');
+                return $this->redirectToRoute('expediente_index');
+            }  
         }
-        $this->addFlash('success', 'Historial eliminado con éxito');
-        return $this->redirectToRoute('historial_familiar_index',['familiar' => $familiar->getId()]);
+
+        if($familiar->getFamiliaresExpedientes()[0]->getExpediente()->getHabilitado()){
+            if ($this->isCsrfTokenValid('delete'.$historialFamiliar->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($historialFamiliar);
+                $entityManager->flush();
+            }
+            $this->addFlash('success', 'Historial eliminado con éxito');
+            return $this->redirectToRoute('historial_familiar_index',['familiar' => $familiar->getId()]);
+        }else{
+            $this->addFlash('fail','Este paciente no está habilitado, para poder hacer uso de el consulte con su superior para habilitar el paciente');
+            return $this->redirectToRoute('expediente_index');
+        }
     }
 }
